# docs/backend-services-elevenlabs.openapi.yaml
openapi: 3.0.3
info:
  title: Ignis Backend Services for ElevenLabs Tools
  description: |
    API used by the ElevenLabs Conversational AI agent during real-time patient intake calls.
    Scope: patient lookup, patient create/update, appointment slots, booking, urgent queue,
    and emergency transfer. Backend is implemented by Next.js API Routes (or equivalent)
    and backed by Aidbox FHIR.
  version: 1.0.0
  contact:
    name: Ignis Team
servers:
  - url: /api
    description: Relative to app origin (e.g. Next.js)
...
paths:
...
  /appointments/slots:
    get:
      operationId: get_available_slots
      summary: Get available slots
      description: |
        Return bookable appointment slots for a given date (and optionally practitioner).
        **When the agent uses it:** When the caller is ready to book (Regular or Urgent
        with same-day slot). Agent asks for preferred date, then calls this tool.
        **FHIR:** Slot resources or derived from Schedule + existing Appointments.
      tags:
        - appointments
      parameters:
        - name: date
          in: query
          description: ISO 8601 date (YYYY-MM-DD)
          required: true
          schema:
            type: string
            format: date
        - name: urgency
          in: query
          description: Filter slots by urgency level. Urgent returns only same-day slots.
          required: false
          schema:
            type: string
            enum:
              - routine
              - urgent
            default: routine
        - name: practitionerId
          in: query
          description: FHIR Practitioner id; if omitted, all practitioners (or default)
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Max slots to return (default 10)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: List of available slots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotsResponse'
        '400':
          description: Invalid date or limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Backend or Aidbox error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
  /appointments:
    post:
      operationId: book_appointment
      summary: Book appointment
      description: |
        Create an Appointment for a chosen slot and patient.
        **When the agent uses it:** After the patient has chosen a slot from the
        offered times. Agent confirms and calls this tool, then repeats confirmation.
        **FHIR:** POST /fhir/Appointment; update Slot status to busy.
      tags:
        - appointments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookAppointmentRequest'
      responses:
        '201':
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookAppointmentResponse'
        '400':
          description: Validation failed (e.g. missing slotId or patientId)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: slotId or patientId not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorWithResource'
        '409':
          description: Slot no longer available (double-book)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotUnavailableError'
        '500':
          description: Backend or Aidbox error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
  /appointments/cancel/{appointmentId}:
    post:
      operationId: cancel_appointment
      summary: Cancel appointment
      description: |
        Cancel an existing appointment.
        **When the agent uses it:** When the patient wants to cancel an appointment.
        First use patient_lookup to find upcoming appointment, then cancel.
        **FHIR:** Update Appointment status to cancelled; free up Slot.
      tags:
        - appointments
      parameters:
        - name: appointmentId
          in: path
          description: FHIR Appointment id
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for cancellation (optional)
      responses:
        '200':
          description: Appointment cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelAppointmentResponse'
        '404':
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '409':
          description: Appointment already cancelled or in the past
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentConflictError'
        '500':
          description: Backend or Aidbox error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
  /appointments/intake-questions:
    get:
      operationId: get_intake_questions
      summary: Get intake questions for appointment
      description: |
        Returns follow-up questions to ask the patient after booking an appointment.
        Questions are tailored based on appointment type and reason.
        **When the agent uses it:** After successfully booking an appointment,
        call this to get relevant intake questions to ask the patient.
        **Example:** After booking a vaccination appointment, ask about allergies.
      tags:
        - appointments
      parameters:
        - name: reason
          in: query
          description: The reason for the appointment (e.g., "Impfung", "Vorsorge", "Rückenschmerzen")
          required: false
          schema:
            type: string
        - name: type
          in: query
          description: Appointment type
          required: false
          schema:
            type: string
            enum:
              - routine
              - urgent
              - followup
            default: routine
        - name: patientId
          in: query
          description: Patient ID (optional, for personalized questions)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of intake questions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntakeQuestionsResponse'
...
components:
  schemas:
...
    Slot:
      type: object
      required:
        - slotId
        - start
        - end
      properties:
        slotId:
          type: string
          description: Id to use in book_appointment request
        start:
          type: string
          format: date-time
          description: ISO 8601 with timezone (e.g. Europe/Berlin)
        end:
          type: string
          format: date-time
        practitionerId:
          type: string
        practitionerDisplay:
          type: string
    SlotsResponse:
      type: object
      required:
        - slots
      properties:
        slots:
          type: array
          items:
            $ref: '#/components/schemas/Slot'

    # --- Book appointment ---
    BookAppointmentRequest:
      type: object
      required:
        - slotId
        - patientId
      properties:
        slotId:
          type: string
          description: Id from get_available_slots
        patientId:
          type: string
          description: FHIR Patient id
        practitionerId:
          type: string
        type:
          type: string
          enum:
            - routine
            - urgent
          description: Maps to appointmentType
        reason:
          type: string
          description: Reason or description (e.g. Impfung)
    BookAppointmentResponse:
      type: object
      required:
        - appointment
        - start
        - end
      properties:
        appointment:
          type: object
          description: FHIR R4 Appointment resource
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        confirmationMessage:
          type: string
          description: Optional pre-written confirmation text for the agent

    # --- Urgent queue ---
...
    CancelAppointmentResponse:
      type: object
      required:
        - cancelled
      properties:
        cancelled:
          type: boolean
          description: true if successfully cancelled
        appointmentId:
          type: string
        message:
          type: string
          description: Confirmation message for the agent
    AppointmentConflictError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: appointment_conflict
        reason:
          type: string
          description: e.g. "already_cancelled", "in_the_past"

    # --- Request callback ---
...
    IntakeQuestionsResponse:
      type: object
      required:
        - questions
        - totalQuestions
      properties:
        questions:
          type: array
          items:
            type: object
            required:
              - id
              - question
              - type
            properties:
              id:
                type: string
                description: Unique question identifier
              question:
                type: string
                description: Question text in German
              question_en:
                type: string
                description: Question text in English
              type:
                type: string
                enum:
                  - text
                  - number
                  - boolean
                description: Expected answer type
              required:
                type: boolean
                description: Whether the question must be answered
        totalQuestions:
          type: integer
          description: Total number of questions
        patientId:
          type: string
          nullable: true
        appointmentType:
          type: string
        reason:
          type: string
          nullable: true

    # --- Common errors ---
    ValidationError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: validation_failed
        message:
          type: string
...
    NotFoundError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: not_found
    NotFoundErrorWithResource:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: not_found
        resource:
          type: string
          description: e.g. slot, patient
    SlotUnavailableError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: slot_unavailable
    InternalError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: internal
        message:
          type: string